Added basic history
Removed the ImgurURL in favor of a NSDictionary for passing information
around in the app. All the other uploaders need to follow this
convention.

Added submenu for history, and populate it when files upload. Stores the
dict for that file in the menu item object itself, thus the
MenuItemWithDict class.