Enable Start at Login
This creates an embedded helper app that will handle the Start at Login
process per Apple's documentaiton.

Apple gives us two options for creating a Login Item

- Adding Login Items Using a Shared File List
- Adding Login Items Using the Service Management Framework

The first is easier and nicer to the user, the latter works with
Sandboxed apps. I am trying to make sure that captured works with a
sandbox.