Use AWS V4 Signing for Request
This was a pretty major refactor of AWS upload:

- Rewrite everything in Swift
- Manually sign the request. Because AWS SDK doesn't support OS X I
  needed to roll my own signing. Two reasons: Newer Amazon data centers
  don't support the old V1 format, I didn't want to pull in a hacked
  together version of the SDK to work with OS X.

The S3V4Signer was written to be as independent of the HTTP lib as
possible. It will just take a few params and create and return the
necessary headers. Care should be taken to make sure none of those
headers are modified after signing.