Refresh expired Imgur Tokens
Not happy about the code organization here, I had to teach the
`ImgurAccount` about the `AccountManager` so that updating the token would
get saved to disk.

Also torn about how I handled the internal ivar for `account` in
`ImgurUploader`. It really should be an `ImgurAccount` and not a generic
`Account` since it needs those methods in a few places. Right now I just
cast it when I need access -- but that feels a bit hacky.