Initial Drag & Drop to the status icon
This is an oft-requested feature.

Right now I wanted to get the mechanics of the upload working. Required
a little refactoring of the `Command` class so I could do an upload
without taking a screen shot.

As a result the GCD calls were refactored out of `Command` and into the
`AppDelegate`. I think I will try to handle all the threading in that
one top level class.