Add "Link Dropbox" account
Does not upload to Dropbox yet, just links the account via OAuth.

There can be only one Dropbox Account associated when using their SDK.
So since I don't want to do all the OAuth and API access manually I am
limited to that. This means that I disable the "Add Dropbox" menu item
under the accounts when there is one already enabled. It also means the
DropboxDetailViewController is full of calls to global functions to link
and unlink.  ¯\_(ツ)_/¯