Add skeleton for dropbox uploader, get annoyed, and disable it
Dropbox has kinda made Captured a feature in their app. So I am thinking
I drop support for this service anyway. It seemed like it was going to
be pretty easy to support, but since their SDK only handles one account
at a time AND couldn't get the `uploadFile` API to work after an hour or
two of swearing.

The menu item is disabled for now, so I am punting on this for the time
being. May want re revisit or rip it out entirely.