Add first-run dialog with a popover
Fixes #36

The popover will point to the status icon in the menubar. It currently
only shows during first run. And is only dismissed when the user clicks
on the "Got it" button.

I could use the EventMonitor class (from Ray Wenderlich's blog) to
listen for global events and dismiss the popover when the user clicks
outside of the view. But was wondering if that overhead was something I
needed for a view that will only run once.

Still todo:

- Add copy to the dialog
- Dismiss the dialog when the user clicks on the status menu icon