Launch At Login Helper App
Fixes #44

Add an embedded helper app to work with the service
management system. This will use `launchctl` to handle running Captured
at login. This is required for Sandbox Apps.

The biggest downside is that the app will no longer appear in the "Login
Items" under system preferences.

I also dropped external dependencies to `StartAtLoginController` and
`LaunchAtLoginController` since they both had deprecation warnings and I
was able to implement a simpler `LoginItem` class that could verify my
helper was registered with Service Management.

Also updated a few of the Carthage dependencies while removing the

A few references:

 - http://martiancraft.com/blog/2015/01/login-items/
 - https://github.com/alexzielenski/StartAtLoginController
 - https://github.com/keith/LoginItemTest

 libraries.