Add option to disable status bar icon
I also comment out and clean up a bunch of cruft.

I also debated how to get the event up to the AppDelegate with the new
setting for EnableMenuBarIcon. Eventually I decided that it was best to
manually control it in the GeneralPreferencesViewController and just
explicitly set it in defaults and call the AppDelegate to re-create the
StatusItem.