Refactor Account Protocol
Remove the `init(Account)` method from the `Uploader` protocol. This lets
us use different constructors for each of the `*Uploader` classes, so we
can require the appropriate type to be passed in when creating the
uploader.

I like how this cleans up the `UploadManager` switch statement, removing
the need for the stringly typed switch. Instead we are able to check the
class of each Account object and call the correct factory.

Also learned about swift's `where` in conditionals. This was handy for
checking multiple things in a single `if` to verify that everything
worked.  Lets me have only one `else` with the error case, instead of a
nested tree where we check for the error on each step.