Add temporary entitlement for `screencapture`
Fixes #48

Due to sandbox rules we can no longer run an NSTask for an executable
that lives outside our app bundle. This adds an entitlement to disable
this check, which will whitelist this app with the OS.

Most interesting, this manifested itself with the following message in
the console:

    screencapture: cannot run two interactive screen captures at a time

I was not running it twice, but that was just a misleading commit
message. Turns out if I ran the Mac App Store version I would get a
crash that was a little more informative:

    SandboxViolation: screencapture(3533) deny mach-register
    com.apple.screencapture.interactive

A few references:

- https://forums.developer.apple.com/thread/51236
- https://eternalstorms.wordpress.com/2016/08/01/screencapture-and-the-sandbox-in-macos-sierra/
- https://eternalstorms.wordpress.com/2016/09/10/deconstructing-and-reimplementing-macos-screencapture-cli/