Add a CONTRIBUTING document
I've been working on this project for nearly a decade. So, now that I am
finally freeing the source I want to make sure to do things as best as
I can. Contribution guide and a CoC help achieve that.

Much of this is based on [Knative's CONTRIBUTING][1] document along with
the [Contributor Covenant CoC][2].

I went with this because @bobcatfish suggested it in her [amazing talk at
Devops Days][3], and overall liked the tone and guidance it provided.

[1]: https://github.com/knative/docs/blob/master/community/CONTRIBUTING.md
[2]: http://contributor-covenant.org/version/1/4/
[3]: https://docs.google.com/presentation/d/1O44ZytTXQ9DleDXgGqtzUXbCkt9czYmuPpbVnRH38XM/edit